+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- fix updating position when only one coordinate changes
+
 ## 1.0.0
 
 - initial release